+++ controllers/serviceController_int_test.go
@@ -114,7 +114,7 @@ var _ = Describe("Creating ingress objects with the ingress generator", func() {
 				return err == nil
 			}, timeoutInterval, pollingInterval).Should(BeTrue())
 
-			Expect(len(expectedIngress.Items)).Should(Equal(0))
+			Expect(expectedIngress.Items).Should(HaveLen(0))
 		})
 
 		It("Should create ingress object for simple webapp service", func() {
@@ -229,9 +229,9 @@ var _ = Describe("Creating ingress objects with the ingress generator", func() {
 			Expect(*expectedIngress.Items[0].Spec.Rules[0].HTTP.Paths[0].PathType).Should(Equal(networking.PathTypePrefix))
 			Expect(expectedIngress.Items[0].Spec.Rules[0].HTTP.Paths[0].Backend.Service.Name).Should(Equal("nexus"))
 			Expect(expectedIngress.Items[0].Spec.Rules[0].HTTP.Paths[0].Backend.Service.Port.Number).Should(Equal(int32(8082)))
-
-			_, ok := expectedIngress.Items[0].Annotations[ingressRewriteTargetAnnotation]
-			Expect(ok).Should(BeFalse())
+			Expect(expectedIngress.Items[0].Annotations).Should(HaveKeyWithValue(ingressConfigurationSnippetAnnotation, "proxy_set_header Accept-Encoding \"identity\";"))
+			Expect(expectedIngress.Items[0].Annotations).Should(HaveKeyWithValue("example-key", "example-value"))
+			Expect(expectedIngress.Items[0].Annotations).Should(Not(HaveKey(ingressRewriteTargetAnnotation)))
 
 			Expect(expectedIngress.Items[1].Namespace).Should(Equal(myNamespace))
 			Expect(expectedIngress.Items[1].Name).Should(Equal("nexus-docker-registry"))
@@ -240,7 +240,9 @@ var _ = Describe("Creating ingress objects with the ingress generator", func() {
 			Expect(*expectedIngress.Items[1].Spec.Rules[0].HTTP.Paths[0].PathType).Should(Equal(networking.PathTypePrefix))
 			Expect(expectedIngress.Items[1].Spec.Rules[0].HTTP.Paths[0].Backend.Service.Name).Should(Equal("nexus"))
 			Expect(expectedIngress.Items[1].Spec.Rules[0].HTTP.Paths[0].Backend.Service.Port.Number).Should(Equal(int32(8082)))
-			Expect(expectedIngress.Items[1].Annotations[ingressRewriteTargetAnnotation]).Should(Equal("/nexus/repository/docker-registry/v2"))
+			Expect(expectedIngress.Items[1].Annotations).Should(HaveKeyWithValue(ingressRewriteTargetAnnotation, "/nexus/repository/docker-registry/v2"))
+			Expect(expectedIngress.Items[1].Annotations).Should(HaveKeyWithValue(ingressConfigurationSnippetAnnotation, "proxy_set_header Accept-Encoding \"identity\";"))
+			Expect(expectedIngress.Items[1].Annotations).Should(HaveKeyWithValue("example-key", "example-value"))
 		})
 
 		It("Should create ssl cert", func() {