+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.11.0] - 2023-04-06
+### Added
+- [#24] Apply additional ingress annotations from dogu service to ingress object.
+
 ## [v0.10.0] - 2023-03-24
 ### Added
 - [#22] Add ssl api to renew the selfsigned certificate of the Cloudogu Ecosystem.